+++ classes/booking.php
@@ -24,6 +24,7 @@
 
 namespace mod_booking;
 
+use cache_helper;
 use course_modinfo;
 use html_writer;
 use local_entities\local\entities\entitydate;
@@ -1555,5 +1556,23 @@ class booking {
         } else {
             return [];
         }
+    }
+
+    /**
+     * Helper function to purge all caches for a booking instance.
+     * @param int $cmid
+     */
+    public static function purge_cache_for_booking_instance_by_cmid(int $cmid) {
+        cache_helper::invalidate_by_event('setbackbookinginstances', [$cmid]);
+        cache_helper::purge_by_event('setbackoptionsettings');
+        cache_helper::purge_by_event('setbackoptionstable');
+        cache_helper::purge_by_event('setbacksemesters');
+
+        // We also need to set back Wunderbyte table cache!
+        cache_helper::purge_by_event('setbackencodedtables');
+        cache_helper::purge_by_event('setbackeventlogtable');
+
+        // Make sure, we destroy singletons too.
+        singleton_service::destroy_booking_singleton_by_cmid($cmid);
     }
 }