+++ lib.php
@@ -686,11 +686,7 @@ function booking_add_instance($booking) {
     booking_grade_item_update($booking);
 
     // When adding an instance, we need to invalidate the cache for booking instances.
-    cache_helper::invalidate_by_event('setbackbookinginstances', [$cmid]);
-
-    // Also purge caches for options table and booking_option_settings.
-    cache_helper::purge_by_event('setbackoptionstable');
-    cache_helper::purge_by_event('setbackoptionsettings');
+    booking::purge_cache_for_booking_instance_by_cmid($cmid);
 
     return $booking->id;
 }
@@ -876,16 +872,7 @@ function booking_update_instance($booking) {
     $event->trigger();
 
     // When updating an instance, we need to invalidate the cache for booking instances.
-    cache_helper::invalidate_by_event('setbackbookinginstances', [$cm->id]);
-
-    // Also purge caches for options table, semesters and booking_option_settings.
-    cache_helper::purge_by_event('setbackoptionstable');
-    cache_helper::purge_by_event('setbacksemesters');
-    cache_helper::purge_by_event('setbackoptionsettings');
-
-    // We also need to set back Wunderbyte table cache!
-    cache_helper::purge_by_event('setbackencodedtables');
-    cache_helper::purge_by_event('setbackeventlogtable');
+    booking::purge_cache_for_booking_instance_by_cmid($cm->id);
 
     return $DB->update_record('booking', $booking);
 }
@@ -1801,11 +1788,7 @@ function booking_delete_instance($id) {
     }
 
     // When deleting an instance, we need to invalidate the cache for booking instances.
-    cache_helper::invalidate_by_event('setbackbookinginstances', [$cm->id]);
-
-    // Also purge caches for options table and booking_option_settings.
-    cache_helper::purge_by_event('setbackoptionstable');
-    cache_helper::purge_by_event('setbackoptionsettings');
+    booking::purge_cache_for_booking_instance_by_cmid($cm->id);
 
     return $result;
 }